[+]Added account namespace for companies. Refactored companies controller and test for it.
