+++ app/views/account/companies/edit.html.slim
@@ -1,7 +1,7 @@
 h1.col-md-6.col-md-offset-3 Edit Company
 .row
   .col-md-3.col-md-offset-3
-    = simple_form_for @company, url: company_path,
+    = simple_form_for @company, url: account_company_path,
                                 defaults: { input_html: { class: 'form-control' } } do |f|
       = f.input :name
       = f.input :domain, hint: "Only '-' of special characters is allowed."
