+++ app/views/account/companies/index.html.slim
@@ -1,0 +1,7 @@
+h1 All companies
+= link_to "Add new company", new_account_company_path, class: 'btn btn-primary'
+
+ul.company
+  - @companies.each do |company|
+    li
+      = link_to company.name, account_company_path(company)
