--- app/views/companies/index.html.slim
@@ -1,7 +1,0 @@
-h1 All companies
-= link_to "Add new company", new_company_path, class: 'btn btn-primary'
-
-ul.company
-  - @companies.each do |company|
-    li
-      = link_to company.name, company_path(company)
