--- app/views/companies/show.html.slim
@@ -1,9 +1,0 @@
-h1= @company.name
-.btn-group
-  = link_to  "Edit", edit_company_path, class: 'btn btn-primary'
-  = link_to "Delete", company_path, method: :delete,\
-                      data: { confirm: 'Are you certain you want to delete this company?' },\
-                      class: 'btn btn-primary'
-
-ul.company
-  li= @company.domain
